agents/pyproject.toml
[project]
name = "bud2-agents"
version = "0.1.0"
description = "bud2 conversational agents service"
requires-python = ">=3.12"
dependencies = [
    "attrs>=22.2.0",
    "fastapi>=0.124.0",
    "google-adk>=1.0.0",
    "google-cloud-aiplatform>=1.128.0",
    "httpx>=0.28.0",
    "psycopg[binary,pool]>=3.3.0",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.12.0",
    "python-dateutil>=2.8.0",
    "structlog>=25.0.0",
    "uvicorn[standard]>=0.38.0",
]

[project.scripts]
bud2-agents = "bud2.main:main"
bud2-agents-migrate = "bud2.infra.postgres.migrate:main"

[dependency-groups]
dev = [
    "mypy>=1.19.0",
    "openapi-python-client>=0.28.0",
    "pytest>=9.0.0",
    "pytest-asyncio>=1.3.0",
    "respx>=0.22.0",
    "ruff>=0.14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bud2"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]
extend-exclude = ["src/bud2/infra/bud2/generated"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.12"
mypy_path = "src"
strict = true
warn_unused_ignores = true
warn_return_any = true
disallow_untyped_defs = true
exclude = ["src/bud2/infra/bud2/generated"]

[[tool.mypy.overrides]]
module = ["google.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["bud2.infra.bud2.generated.*"]
ignore_errors = true
